Inline AbstractClass test data into test
