--- Source/Moq.Analyzers.Test/Data/AbstractClass.MismatchArgs.cs
@@ -1,34 +1,0 @@
-﻿namespace Moq.Analyzers.Test.Data.AbstractClass.MismatchArgs;
-
-internal abstract class AbstractClassDefaultCtor
-{
-    protected AbstractClassDefaultCtor()
-    {
-    }
-}
-
-internal abstract class AbstractClassWithCtor
-{
-    protected AbstractClassWithCtor(int a)
-    {
-    }
-
-    protected AbstractClassWithCtor(int a, string b)
-    {
-    }
-}
-
-internal class MyUnitTests
-{
-    private void TestBad()
-    {
-        // The class has a ctor that takes an Int32 but passes a String
-        var mock = new Mock<AbstractClassWithCtor>("42");
-
-        // The class has a ctor with two arguments [Int32, String], but they are passed in reverse order
-        var mock1 = new Mock<AbstractClassWithCtor>("42", 42);
-
-        // The class has a ctor but does not take any arguments
-        var mock2 = new Mock<AbstractClassDefaultCtor>(42);
-    }
-}
